requirements.txt
tweepy
Pillow
requests
bs4
os
dotenv
